backend/requirements.txt
fastapi==0.115.0
uvicorn==0.32.1
python-multipart==0.0.9
pydantic==2.11.7
python-dotenv==1.0.1
aiofiles==23.2.1
boto3==1.34.34
supabase==2.27.1
httpx==0.27.2
openpyxl==3.1.2
groq==0.4.2
openai==1.14.0
deepgram-sdk==3.7.0
tenacity==9.0.0
google-genai>=0.2.0
google-cloud-speech==2.27.0
speechmatics-batch==0.4.4